[qa] move Schedule as a widget component
