+++ src/components/pages/Asset.vue
@@ -351,25 +351,25 @@ import { sortByName } from '@/lib/sorting'
 import { entityMixin } from '@/components/mixins/entity'
 import { formatListMixin } from '@/components/mixins/format'
 
-import ButtonSimple from '@/components/widgets/ButtonSimple'
-import ConceptCard from '@/components/widgets/ConceptCard'
-import ComboboxNumber from '@/components/widgets/ComboboxNumber'
-import ComboboxStatus from '@/components/widgets/ComboboxStatus'
-import DescriptionCell from '@/components/cells/DescriptionCell'
-import EditAssetModal from '@/components/modals/EditAssetModal'
-import EntityChat from '@/components/pages/entities/EntityChat'
-import EntityNews from '@/components/pages/entities/EntityNews'
-import EntityOutputFiles from '@/components/pages/entities/EntityOutputFiles'
-import EntityPreviewFiles from '@/components/pages/entities/EntityPreviewFiles'
-import EntityTimeLogs from '@/components/pages/entities/EntityTimeLogs'
-import EntityTaskList from '@/components/lists/EntityTaskList'
-import EntityThumbnail from '@/components/widgets/EntityThumbnail'
-import PageSubtitle from '@/components/widgets/PageSubtitle'
-import RouteSectionTabs from '@/components/widgets/RouteSectionTabs'
-import Schedule from '@/components/pages/schedule/Schedule'
-import TableInfo from '@/components/widgets/TableInfo'
-import TaskTypeName from '@/components/widgets/TaskTypeName'
-import TaskInfo from '@/components/sides/TaskInfo'
+import ButtonSimple from '@/components/widgets/ButtonSimple.vue'
+import ConceptCard from '@/components/widgets/ConceptCard.vue'
+import ComboboxNumber from '@/components/widgets/ComboboxNumber.vue'
+import ComboboxStatus from '@/components/widgets/ComboboxStatus.vue'
+import DescriptionCell from '@/components/cells/DescriptionCell.vue'
+import EditAssetModal from '@/components/modals/EditAssetModal.vue'
+import EntityChat from '@/components/pages/entities/EntityChat.vue'
+import EntityNews from '@/components/pages/entities/EntityNews.vue'
+import EntityOutputFiles from '@/components/pages/entities/EntityOutputFiles.vue'
+import EntityPreviewFiles from '@/components/pages/entities/EntityPreviewFiles.vue'
+import EntityTimeLogs from '@/components/pages/entities/EntityTimeLogs.vue'
+import EntityTaskList from '@/components/lists/EntityTaskList.vue'
+import EntityThumbnail from '@/components/widgets/EntityThumbnail.vue'
+import PageSubtitle from '@/components/widgets/PageSubtitle.vue'
+import RouteSectionTabs from '@/components/widgets/RouteSectionTabs.vue'
+import Schedule from '@/components/widgets/Schedule.vue'
+import TableInfo from '@/components/widgets/TableInfo.vue'
+import TaskTypeName from '@/components/widgets/TaskTypeName.vue'
+import TaskInfo from '@/components/sides/TaskInfo.vue'
 
 export default {
   name: 'asset',
