+++ src/components/pages/Edit.vue
@@ -596,26 +596,26 @@ import { previewRoomMixin } from '@/components/mixins/previewRoom'
 import { playerMixin } from '@/components/mixins/player'
 import { DEFAULT_NB_FRAMES_PICTURE } from '@/lib/playlist'
 
-import ButtonSimple from '@/components/widgets/ButtonSimple'
-import ComboboxNumber from '@/components/widgets/ComboboxNumber'
-import DescriptionCell from '@/components/cells/DescriptionCell'
-import EditEditModal from '@/components/modals/EditEditModal'
-import EntityNews from '@/components/pages/entities/EntityNews'
-import EntityPreviewFiles from '@/components/pages/entities/EntityPreviewFiles'
-import EntityTaskList from '@/components/lists/EntityTaskList'
-import EntityTimeLogs from '@/components/pages/entities/EntityTimeLogs'
-import EntityThumbnail from '@/components/widgets/EntityThumbnail'
-import ColorPicker from '@/components/widgets/ColorPicker'
-import ComboboxStyled from '@/components/widgets/ComboboxStyled'
-import PageSubtitle from '@/components/widgets/PageSubtitle'
-import PencilPicker from '@/components/widgets/PencilPicker'
-import PreviewRoom from '@/components/widgets/PreviewRoom'
-import PreviewsPerTaskType from '@/components/previews/PreviewsPerTaskType'
-import RawVideoPlayer from '@/components/pages/playlists/RawVideoPlayer'
-import Schedule from '@/components/pages/schedule/Schedule'
-import Spinner from '@/components/widgets/Spinner'
-import TaskInfo from '@/components/sides/TaskInfo'
-import VideoProgress from '@/components/previews/VideoProgress'
+import ButtonSimple from '@/components/widgets/ButtonSimple.vue'
+import ComboboxNumber from '@/components/widgets/ComboboxNumber.vue'
+import DescriptionCell from '@/components/cells/DescriptionCell.vue'
+import EditEditModal from '@/components/modals/EditEditModal.vue'
+import EntityNews from '@/components/pages/entities/EntityNews.vue'
+import EntityPreviewFiles from '@/components/pages/entities/EntityPreviewFiles.vue'
+import EntityTaskList from '@/components/lists/EntityTaskList.vue'
+import EntityTimeLogs from '@/components/pages/entities/EntityTimeLogs.vue'
+import EntityThumbnail from '@/components/widgets/EntityThumbnail.vue'
+import ColorPicker from '@/components/widgets/ColorPicker.vue'
+import ComboboxStyled from '@/components/widgets/ComboboxStyled.vue'
+import PageSubtitle from '@/components/widgets/PageSubtitle.vue'
+import PencilPicker from '@/components/widgets/PencilPicker.vue'
+import PreviewRoom from '@/components/widgets/PreviewRoom.vue'
+import PreviewsPerTaskType from '@/components/previews/PreviewsPerTaskType.vue'
+import RawVideoPlayer from '@/components/pages/playlists/RawVideoPlayer.vue'
+import Schedule from '@/components/widgets/Schedule.vue'
+import Spinner from '@/components/widgets/Spinner.vue'
+import TaskInfo from '@/components/sides/TaskInfo.vue'
+import VideoProgress from '@/components/previews/VideoProgress.vue'
 
 export default {
   name: 'edit',
