+++ src/components/pages/Episode.vue
@@ -246,23 +246,23 @@ import { episodifyRoute } from '@/lib/path'
 import { entityMixin } from '@/components/mixins/entity'
 import { formatListMixin } from '@/components/mixins/format'
 
-import ButtonSimple from '@/components/widgets/ButtonSimple'
-import ComboboxNumber from '@/components/widgets/ComboboxNumber'
-import DescriptionCell from '@/components/cells/DescriptionCell'
-import EditEpisodeModal from '@/components/modals/EditEpisodeModal'
-import EntityChat from '@/components/pages/entities/EntityChat'
-import EntityNews from '@/components/pages/entities/EntityNews'
-import EntityPreviewFiles from '@/components/pages/entities/EntityPreviewFiles'
-import EntityTaskList from '@/components/lists/EntityTaskList'
-import EntityTimeLogs from '@/components/pages/entities/EntityTimeLogs'
-import EntityThumbnail from '@/components/widgets/EntityThumbnail'
-import PageTitle from '@/components/widgets/PageTitle'
-import PageSubtitle from '@/components/widgets/PageSubtitle'
-import RouteSectionTabs from '@/components/widgets/RouteSectionTabs'
-import Schedule from '@/components/pages/schedule/Schedule'
-import TableInfo from '@/components/widgets/TableInfo'
-import TaskInfo from '@/components/sides/TaskInfo'
-import TaskTypeName from '@/components/widgets/TaskTypeName'
+import ButtonSimple from '@/components/widgets/ButtonSimple.vue'
+import ComboboxNumber from '@/components/widgets/ComboboxNumber.vue'
+import DescriptionCell from '@/components/cells/DescriptionCell.vue'
+import EditEpisodeModal from '@/components/modals/EditEpisodeModal.vue'
+import EntityChat from '@/components/pages/entities/EntityChat.vue'
+import EntityNews from '@/components/pages/entities/EntityNews.vue'
+import EntityPreviewFiles from '@/components/pages/entities/EntityPreviewFiles.vue'
+import EntityTaskList from '@/components/lists/EntityTaskList.vue'
+import EntityTimeLogs from '@/components/pages/entities/EntityTimeLogs.vue'
+import EntityThumbnail from '@/components/widgets/EntityThumbnail.vue'
+import PageTitle from '@/components/widgets/PageTitle.vue'
+import PageSubtitle from '@/components/widgets/PageSubtitle.vue'
+import RouteSectionTabs from '@/components/widgets/RouteSectionTabs.vue'
+import Schedule from '@/components/widgets/Schedule.vue'
+import TableInfo from '@/components/widgets/TableInfo.vue'
+import TaskInfo from '@/components/sides/TaskInfo.vue'
+import TaskTypeName from '@/components/widgets/TaskTypeName.vue'
 
 export default {
   name: 'episode',
