+++ src/components/pages/Person.vue
@@ -187,7 +187,7 @@ import ComboboxProduction from '@/components/widgets/ComboboxProduction.vue'
 import KanbanBoard from '@/components/lists/KanbanBoard.vue'
 import PeopleAvatar from '@/components/widgets/PeopleAvatar.vue'
 import RouteSectionTabs from '@/components/widgets/RouteSectionTabs.vue'
-import Schedule from '@/components/pages/schedule/Schedule.vue'
+import Schedule from '@/components/widgets/Schedule.vue'
 import SearchField from '@/components/widgets/SearchField.vue'
 import SearchQueryList from '@/components/widgets/SearchQueryList.vue'
 import TimesheetList from '@/components/lists/TimesheetList.vue'
