+++ src/components/widgets/Schedule.vue
@@ -1800,25 +1800,25 @@ const setItemPositions = (items, attributeName, unitOfTime = 'days') => {
 
   .schedule.zoom-level-0 {
     .timeline-content {
-      background-image: url('../../../assets/background/schedule-dark-1.png');
+      background-image: url('@/assets/background/schedule-dark-1.png');
     }
   }
 
   .schedule.zoom-level-1 {
     .timeline-content {
-      background-image: url('../../../assets/background/schedule-dark-1-weekend.png');
+      background-image: url('@/assets/background/schedule-dark-1-weekend.png');
     }
   }
 
   .schedule.zoom-level-2 {
     .timeline-content {
-      background-image: url('../../../assets/background/schedule-dark-2-weekend.png');
+      background-image: url('@/assets/background/schedule-dark-2-weekend.png');
     }
   }
 
   .schedule.zoom-level-3 {
     .timeline-content {
-      background-image: url('../../../assets/background/schedule-dark-3-weekend.png');
+      background-image: url('@/assets/background/schedule-dark-3-weekend.png');
     }
   }
 
@@ -2180,7 +2180,7 @@ const setItemPositions = (items, attributeName, unitOfTime = 'days') => {
 
 .zoom-level-0 {
   .timeline-content {
-    background-image: url('../../../assets/background/schedule-white-1.png');
+    background-image: url('@/assets/background/schedule-white-1.png');
   }
   .day {
     width: 20px;
@@ -2192,7 +2192,7 @@ const setItemPositions = (items, attributeName, unitOfTime = 'days') => {
 
 .zoom-level-1 {
   .timeline-content {
-    background-image: url('../../../assets/background/schedule-white-1-weekend.png');
+    background-image: url('@/assets/background/schedule-white-1-weekend.png');
   }
   .day {
     width: 20px;
@@ -2204,7 +2204,7 @@ const setItemPositions = (items, attributeName, unitOfTime = 'days') => {
 
 .schedule.zoom-level-2 {
   .timeline-content {
-    background-image: url('../../../assets/background/schedule-white-2-weekend.png');
+    background-image: url('@/assets/background/schedule-white-2-weekend.png');
   }
   .day {
     width: 40px;
@@ -2216,7 +2216,7 @@ const setItemPositions = (items, attributeName, unitOfTime = 'days') => {
 
 .schedule.zoom-level-3 {
   .timeline-content {
-    background-image: url('../../../assets/background/schedule-white-3-weekend.png');
+    background-image: url('@/assets/background/schedule-white-3-weekend.png');
   }
   .day {
     width: 60px;
